# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "X-Realty Estate API",
        "description": "API endpoints for real estate management",
        "contact": {
            "name": "X-Realty Estate API Team"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/reaxml/generate/{id}": {
            "get": {
                "tags": [
                    "REAXML"
                ],
                "summary": "Generate REAXML XML for a property",
                "operationId": "784d0575e25e538b3d620ef1f49a1090",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "REAXML XML generated successfully",
                        "content": {
                            "application/xml": {}
                        }
                    },
                    "422": {
                        "description": "Validation error or missing required fields",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}